pyproject.toml
[tool.poetry]
name = "hands-on-aws-cdk-python-exercises"
version = "0.1.0"
description = ""
authors = ["Ben de Haan <2851705-bdehaan@users.noreply.gitlab.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
poetry-plugin-export = "^1.6.0"
python-dotenv = "^1.0.0"
aws-cdk-lib ="^2.121.1"
constructs ="10.0.0,<11.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[tool.bandit]
exclude_dirs = ['tests']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
